+++ src/main/java/com/gulbi/Backend/infrastructure/recommendation/jsonpath/LokiQueryHandler.java
@@ -1,9 +1,8 @@
-package com.gulbi.Backend.domain.rental.recommandation.service.query;
+package com.gulbi.Backend.infrastructure.recommendation.jsonpath;
 
-import static com.gulbi.Backend.domain.rental.recommandation.service.query.JsonPathQuery.*;
+import static com.gulbi.Backend.infrastructure.recommendation.jsonpath.JsonPathQuery.*;
 
-import com.gulbi.Backend.domain.rental.recommandation.dto.CategoryPair;
-import com.gulbi.Backend.domain.rental.recommandation.vo.PriorityCategoriesMap;
+import com.gulbi.Backend.domain.rental.recommandation.service.query.QueryHandler;
 import com.gulbi.Backend.domain.rental.recommandation.vo.PopularProductIds;
 import com.gulbi.Backend.domain.rental.recommandation.vo.PriorityCategoriesQueue;
 import com.jayway.jsonpath.JsonPath;
@@ -15,7 +14,7 @@ import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
 @Service
-public class LokiQueryHandler implements QueryHandler{
+public class LokiQueryHandler implements QueryHandler {
     private final String className = this.getClass().getName();
 
     @Override
@@ -28,14 +27,14 @@ public class LokiQueryHandler implements QueryHandler{
 
     @Override
     public PriorityCategoriesQueue getPriorityCategoriesMap(String queryResult) {
-        List<String> countList = JsonPath.read(queryResult, TESt_QUERY.getQuery());
+        List<String> countList = JsonPath.read(queryResult, EXTRACT_TOTAL_RESULT_COUNT.getQuery());
         int count = countList.size() - 1; // 마지막 인덱스부터 시작
         PriorityCategoriesQueue priorityCategoriesQueue = new PriorityCategoriesQueue();
 
         for (; count >= 0; count--) { // 0까지 감소
-            String testQuery = String.format(TEST_QUERY_2.getQuery(), count);
-            String testQuery2 = String.format(TEST_QUERY_3.getQuery(), count);
-            String testQuery3 = String.format(TEST_QUERY_4.getQuery(), count,1);
+            String testQuery = String.format(EXTRACT_BIG_CATEGORY_ID.getQuery(), count);
+            String testQuery2 = String.format(EXTRACT_MIDDLE_CATEGORY_ID.getQuery(), count);
+            String testQuery3 = String.format(EXTRACT_TOTAL_VIEWED.getQuery(), count,1);
 
             String bigCategoryId = JsonPath.read(queryResult,testQuery);
             String midCategoryId = JsonPath.read(queryResult,testQuery2);
